Moved game save compare related code into a GameSaveCompare namespace out from OpenLoco.cpp
