+++ src/OpenLoco/CMakeLists.txt
@@ -136,6 +136,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Vehicles/VehicleSell.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Vehicles/VehicleSpeedControl.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/GameCommands.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameSaveCompare.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameState.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/Colour.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/Gfx.cpp"
@@ -458,6 +459,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/Vehicles/VehicleSpeedControl.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameCommands/GameCommands.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameException.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/GameSaveCompare.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/GameState.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/Colour.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/Gfx.h"
